refactor: reorder e2e controller tests
