+++ src/api/research-study/ResearchStudyModule.ts
@@ -1,13 +1,13 @@
 import { Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 
-import { GetOneReasearchStudyController } from './controllers/GetOneReasearchStudyController'
+import { GetOneResearchStudyController } from './controllers/GetOneResearchStudyController'
 import { SearchResearchStudyController } from './controllers/SearchResearchStudyController'
 import { EsResearchStudyRepository } from './gateways/EsResearchStudyRepository'
 import { ElasticsearchModule } from '../../shared/elasticsearch/ElasticsearchModule'
 
 @Module({
-  controllers: [GetOneReasearchStudyController, SearchResearchStudyController],
+  controllers: [GetOneResearchStudyController, SearchResearchStudyController],
   imports: [ConfigModule, ElasticsearchModule],
   providers: [EsResearchStudyRepository],
 })
